pridavam

+++ src/main/java/com/example/app/Objects/Borrowing.java
@@ -1,13 +1,13 @@
 package com.example.app.Objects;
 
 public class Borrowing {
-    private int bookId;
+    private long bookId;
     private int customerId;
     public long id;
     public Customer user;
     public Book book;
 
-    public int getBookId(){
+    public long getBookId(){
         return this.bookId;
     }
     public int getCustomerId(){
